🔥 (api): remove v2 scoring completed
and rename v3 scoring
