+++ api/src/certification/evaluation/application/jobs/certification-completed-job-controller.js
@@ -1,6 +1,4 @@
 import { JobController } from '../../../../shared/application/jobs/job-controller.js';
-import { AlgorithmEngineVersion } from '../../../shared/domain/models/AlgorithmEngineVersion.js';
-import * as certificationAssessmentRepository from '../../../shared/infrastructure/repositories/certification-assessment-repository.js';
 import { CertificationCompletedJob } from '../../domain/events/CertificationCompleted.js';
 import { usecases } from '../../domain/usecases/index.js';
 
@@ -9,15 +7,8 @@ export class CertificationCompletedJobController extends JobController {
     super(CertificationCompletedJob.name);
   }
 
-  async handle({ data, dependencies = { certificationAssessmentRepository } }) {
+  async handle({ data }) {
     const { assessmentId, locale } = data;
-    const { certificationAssessmentRepository } = dependencies;
-    const certificationAssessment = await certificationAssessmentRepository.get(assessmentId);
-
-    if (AlgorithmEngineVersion.isV3(certificationAssessment.version)) {
-      await usecases.scoreCompletedV3Certification({ certificationAssessment, locale });
-    } else {
-      await usecases.scoreCompletedV2Certification({ certificationAssessment });
-    }
+    await usecases.scoreCompletedCertification({ assessmentId, locale });
   }
 }
